Move concat_tracks and hide reset_track_id
