+++ docs/api/huracanpy.rst
@@ -13,4 +13,3 @@ huracanpy
     trackswhere
     interp_time
     concat_tracks
-    reset_track_id
